+++ packages/storage/src/Storage.ts
@@ -11,23 +11,13 @@
  * and limitations under the License.
  */
 
-import {
-    AWS,
-    ConsoleLogger as Logger,
-    Hub,
-    Credentials
-} from '@aws-amplify/core';
-import * as S3 from 'aws-sdk/clients/s3';
-import { StorageOptions } from './types';
+import { ConsoleLogger as Logger, Parser } from '@aws-amplify/core';
+import AWSS3Provider from './Providers/AWSS3Provider';
+import { StorageProvider } from './types';
 
 const logger = new Logger('StorageClass');
 
-const dispatchStorageEvent = (track, attrs, metrics) => {
-    if (track) {
-        Hub.dispatch('storage', { attrs, metrics }, 'Storage');
-    }
-};
-
+const DEFAULT_PROVIDER = 'AWSS3';
 /**
  * Provide storage methods to use AWS S3
  */
@@ -35,7 +25,9 @@ export default class StorageClass {
     /**
      * @private
      */
-    private _options;
+    private _config;
+    private _pluggables: StorageProvider[];
+    
 
     /**
      * @public
@@ -43,12 +35,18 @@ export default class StorageClass {
     public vault: StorageClass;
 
     /**
-     * Initialize Storage with AWS configurations
-     * @param {Object} options - Configuration object for storage
+     * Initialize Storage
+     * @param {Object} config - Configuration object for storage
      */
-    constructor(options: StorageOptions) {
-        this._options = options;
-        logger.debug('Storage Options', this._options);
+    constructor() {
+        this._config = {};
+        this._pluggables = [];
+        logger.debug('Storage Options', this._config);
+
+        this.get = this.get.bind(this);
+        this.put = this.put.bind(this);
+        this.remove = this.remove.bind(this);
+        this.list = this.list.bind(this);
     }
 
     public getModuleName() {
@@ -56,311 +54,138 @@ export default class StorageClass {
     }
 
     /**
-     * Configure Storage part with aws configuration
-     * @param {Object} config - Configuration of the Storage
+     * add plugin into Storage category
+     * @param {Object} pluggable - an instance of the plugin
+     */
+    public addPluggable(pluggable: StorageProvider) {
+        if (pluggable && pluggable.getCategory() === 'Storage') {
+            this._pluggables.push(pluggable);
+            let config = {};
+            // for backward compatibility
+            if (pluggable.getProviderName() === DEFAULT_PROVIDER && !this._config[DEFAULT_PROVIDER]) {
+                config = pluggable.configure(this._config);  
+            } else {
+                config = pluggable.configure(this._config[pluggable.getProviderName()]);
+            }
+            return config;
+        }
+    }
+
+    /**
+     * Get the plugin object
+     * @param providerName - the name of the plugin 
+     */
+    public getPluggable(providerName: string) {
+        const pluggable = this._pluggables.find(pluggable => pluggable.getProviderName() === providerName);
+        if(pluggable === undefined) {
+            logger.debug('No plugin found with providerName', providerName);
+            return null;
+        } else 
+            return pluggable;
+    }
+
+    /**
+     * Remove the plugin object
+     * @param providerName - the name of the plugin
+     */
+    public removePluggable(providerName:string) {
+        this._pluggables = this._pluggables.filter(pluggable=> pluggable.getProviderName() !== providerName);
+        return;
+    }
+
+    /**
+     * Configure Storage
+     * @param {Object} config - Configuration object for storage
      * @return {Object} - Current configuration
      */
-    configure(options?) {
+    configure(config?) {
         logger.debug('configure Storage');
-        let opt = options ? options.Storage || options : {};
+        if (!config) return this._config;
+        const amplifyConfig = Parser.parseMobilehubConfig(config);
+        this._config = Object.assign({}, this._config, amplifyConfig.Storage);
+        if (!this._config.bucket) { logger.debug('Do not have bucket yet'); }
+        
+        this._pluggables.forEach((pluggable) => {
+            // for backward compatibility
+            if (pluggable.getProviderName() === DEFAULT_PROVIDER && !this._config[DEFAULT_PROVIDER]) {
+                pluggable.configure(this._config);  
+            } else {
+                pluggable.configure(this._config[pluggable.getProviderName()]);
+            }
+        });
 
-        if (options['aws_user_files_s3_bucket']) {
-            opt = {
-                bucket: options['aws_user_files_s3_bucket'],
-                region: options['aws_user_files_s3_bucket_region']
-            };
+        if (this._pluggables.length === 0) {
+            this.addPluggable(new AWSS3Provider());
         }
-        this._options = Object.assign({}, this._options, opt);
-        if (!this._options.bucket) { logger.debug('Do not have bucket yet'); }
 
-        return this._options;
+        return this._config;
     }
 
     /**
     * Get a presigned URL of the file or the object data when download:true
     *
     * @param {String} key - key of the object
-    * @param {Object} [options] - { level : private|protected|public, download: true|false }
-    * @return - A promise resolves to Amazon S3 presigned URL on success
+    * @param {Object} [config] - { level : private|protected|public, download: true|false }
+    * @return - A promise resolves to either a presigned url or the object
     */
-    public async get(key: string, options?): Promise<String|Object> {
-        const credentialsOK = await this._ensureCredentials();
-        if (!credentialsOK) { return Promise.reject('No credentials'); }
-
-        const opt = Object.assign({}, this._options, options);
-        const { bucket, region, credentials, level, download, track, expires } = opt;
-
-        const prefix = this._prefix(opt);
-        const final_key = prefix + key;
-        const s3 = this._createS3(opt);
-        logger.debug('get ' + key + ' from ' + final_key);
-
-        const params: any = {
-            Bucket: bucket,
-            Key: final_key
-        };
-
-        if (download === true) {
-            return new Promise<any>((res, rej) => {
-                s3.getObject(params, (err, data) => {
-                    if (err) {
-                        dispatchStorageEvent(
-                            track,
-                            { method: 'get', result: 'failed' },
-                            null);
-                        rej(err);
-                    } else {
-                        dispatchStorageEvent(
-                            track,
-                            { method: 'get', result: 'success' },
-                            { fileSize: Number(data.Body['length']) });
-                        res(data);
-                    }
-                });
-            });
+    public async get(key: string, config?): Promise<String|Object> {
+        
+        const { provider = DEFAULT_PROVIDER } = config.provider || {};
+        const prov = this._pluggables.find(pluggable => pluggable.getProviderName() === provider);
+        if(prov === undefined) {
+            logger.debug('No plugin found with providerName', provider);
+            Promise.reject('No plugin found in Storage for the provider');
         }
-
-        if (expires) { params.Expires = expires; }
-
-        return new Promise<string>((res, rej) => {
-            try {
-                const url = s3.getSignedUrl('getObject', params);
-                dispatchStorageEvent(
-                    track,
-                    { method: 'get', result: 'success' },
-                    null);
-                res(url);
-            } catch (e) {
-                logger.warn('get signed url error', e);
-                dispatchStorageEvent(
-                    track,
-                    { method: 'get', result: 'failed' },
-                    null);
-                rej(e);
-            }
-        });
+        return prov.get(key, config);
     }
 
     /**
-     * Put a file in S3 bucket specified to configure method
+     * Put a file in storage bucket specified to configure method
      * @param {Stirng} key - key of the object
-     * @param {Object} object - File to be put in Amazon S3 bucket
-     * @param {Object} [options] - { level : private|protected|public, contentType: MIME Types,
+     * @param {Object} object - File to be put in bucket
+     * @param {Object} [config] - { level : private|protected|public, contentType: MIME Types,
      *  progressCallback: function }
      * @return - promise resolves to object on success
      */
-    public async put(key: string, object, options?): Promise<Object> {
-        const credentialsOK = await this._ensureCredentials();
-        if (!credentialsOK) { return Promise.reject('No credentials'); }
-
-        const opt = Object.assign({}, this._options, options);
-        const { bucket, region, credentials, level, track, progressCallback } = opt;
-        const { contentType, contentDisposition, cacheControl, expires, metadata } = opt;
-        const type = contentType ? contentType : 'binary/octet-stream';
-
-        const prefix = this._prefix(opt);
-        const final_key = prefix + key;
-        const s3 = this._createS3(opt);
-        logger.debug('put ' + key + ' to ' + final_key);
-
-        const params: any = {
-            Bucket: bucket,
-            Key: final_key,
-            Body: object,
-            ContentType: type
-        };
-        if (cacheControl) { params.CacheControl = cacheControl; }
-        if (contentDisposition) { params.ContentDisposition = contentDisposition; }
-        if (expires) { params.Expires = expires; }
-        if (metadata) { params.Metadata = metadata; }
-
-        try {
-            const upload = s3
-                .upload(params)
-                .on('httpUploadProgress', progress => {
-                    if (progressCallback) {
-                        if (typeof progressCallback === 'function') {
-                            progressCallback(progress);
-                        } else {
-                            logger.warn('progressCallback should be a function, not a ' + typeof progressCallback);
-                        }
-                    }
-                });
-            const data = await upload.promise();
-            
-            logger.debug('upload result', data);
-            dispatchStorageEvent(
-                track,
-                { method: 'put', result: 'success' },
-                null);
-            
-            return {
-                key: data.Key.substr(prefix.length)
-            };
-        } catch (e) {
-            logger.warn("error uploading", e);
-            dispatchStorageEvent(
-                track,
-                { method: 'put', result: 'failed' },
-                null);
-            
-            throw e;
+    public async put(key: string, object, config?):Promise<Object> { 
+        const { provider = DEFAULT_PROVIDER } = config.provider || {};
+        const prov = this._pluggables.find(pluggable => pluggable.getProviderName() === provider);
+        if(prov === undefined) {
+            logger.debug('No plugin found with providerName', provider);
+            Promise.reject('No plugin found in Storage for the provider');
         }
+        return prov.put(key,object,config);  
     }
 
     /**
      * Remove the object for specified key
      * @param {String} key - key of the object
-     * @param {Object} [options] - { level : private|protected|public }
+     * @param {Object} [config] - { level : private|protected|public }
      * @return - Promise resolves upon successful removal of the object
      */
-    public async remove(key: string, options?): Promise<any> {
-        const credentialsOK = await this._ensureCredentials();
-        if (!credentialsOK) { return Promise.reject('No credentials'); }
-
-        const opt = Object.assign({}, this._options, options, );
-        const { bucket, region, credentials, level, track } = opt;
-
-        const prefix = this._prefix(opt);
-        const final_key = prefix + key;
-        const s3 = this._createS3(opt);
-        logger.debug('remove ' + key + ' from ' + final_key);
-
-        const params = {
-            Bucket: bucket,
-            Key: final_key
-        };
-
-        return new Promise<any>((res, rej) => {
-            s3.deleteObject(params, (err, data) => {
-                if (err) {
-                    dispatchStorageEvent(
-                        track,
-                        { method: 'remove', result: 'failed' },
-                        null);
-                    rej(err);
-                } else {
-                    dispatchStorageEvent(
-                        track,
-                        { method: 'remove', result: 'success' },
-                        null);
-                    res(data);
-                }
-            });
-        });
+    public async remove(key: string, config?): Promise<any> {
+        const { provider = DEFAULT_PROVIDER } = config.provider || {};
+        const prov = this._pluggables.find(pluggable => pluggable.getProviderName() === provider);
+        if(prov === undefined) {
+            logger.debug('No plugin found with providerName', provider);
+            Promise.reject('No plugin found in Storage for the provider');
+        }
+        return prov.remove(key,config);
     }
 
     /**
      * List bucket objects relative to the level and prefix specified
      * @param {String} path - the path that contains objects
-     * @param {Object} [options] - { level : private|protected|public }
+     * @param {Object} [config] - { level : private|protected|public }
      * @return - Promise resolves to list of keys for all objects in path
      */
-    public async list(path, options?): Promise<any> {
-        const credentialsOK = await this._ensureCredentials();
-        if (!credentialsOK) { return Promise.reject('No credentials'); }
-
-        const opt = Object.assign({}, this._options, options);
-        const { bucket, region, credentials, level, download, track } = opt;
-
-        const prefix = this._prefix(opt);
-        const final_path = prefix + path;
-        const s3 = this._createS3(opt);
-        logger.debug('list ' + path + ' from ' + final_path);
-
-        const params = {
-            Bucket: bucket,
-            Prefix: final_path
-        };
-
-        return new Promise<any>((res, rej) => {
-            s3.listObjects(params, (err, data) => {
-                if (err) {
-                    logger.warn('list error', err);
-                    dispatchStorageEvent(
-                        track,
-                        { method: 'list', result: 'failed' },
-                        null);
-                    rej(err);
-                } else {
-                    const list = data.Contents.map(item => {
-                        return {
-                            key: item.Key.substr(prefix.length),
-                            eTag: item.ETag,
-                            lastModified: item.LastModified,
-                            size: item.Size
-                        };
-                    });
-                    dispatchStorageEvent(
-                        track,
-                        { method: 'list', result: 'success' },
-                        null);
-                    logger.debug('list', list);
-                    res(list);
-                }
-            });
-        });
-    }
-
-    /**
-     * @private
-     */
-    _ensureCredentials() {
-        // commented
-        // will cause bug if another user logged in without refreshing page
-        // if (this._options.credentials) { return Promise.resolve(true); }
-
-        return Credentials.get()
-            .then(credentials => {
-                if (!credentials) return false;
-                const cred = Credentials.shear(credentials);
-                logger.debug('set credentials for storage', cred);
-                this._options.credentials = cred;
-
-                return true;
-            })
-            .catch(err => {
-                logger.warn('ensure credentials error', err);
-                return false;
-            });
-    }
-
-    /**
-     * @private
-     */
-    private _prefix(options) {
-        const { credentials, level } = options;
-
-        const customPrefix = options.customPrefix || {};
-        const identityId = options.identityId || credentials.identityId;
-        const privatePath = (customPrefix.private !== undefined ? customPrefix.private : 'private/') + identityId + '/';
-        const protectedPath = (customPrefix.protected !== undefined ?
-            customPrefix.protected : 'protected/') + identityId + '/';
-        const publicPath = customPrefix.public !== undefined ? customPrefix.public : 'public/';
-
-        switch (level) {
-            case 'private':
-                return privatePath;
-            case 'protected':
-                return protectedPath;
-            default:
-                return publicPath;
+    public async list(path, config?): Promise<any> {
+        const { provider = DEFAULT_PROVIDER } = config.provider || {};
+        const prov = this._pluggables.find(pluggable => pluggable.getProviderName() === provider);
+        if(prov === undefined) {
+            logger.debug('No plugin found with providerName', provider);
+            Promise.reject('No plugin found in Storage for the provider');
         }
-    }
-
-    /**
-     * @private
-     */
-    private _createS3(options) {
-        const { bucket, region, credentials } = options;
-        AWS.config.update({
-            region,
-            credentials
-        });
-        return new S3({
-            apiVersion: '2006-03-01',
-            params: { Bucket: bucket },
-            region
-        });
+        return prov.list(path,config); 
     }
 }