+++ packages/aws-amplify-react-native/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.9-unstable.2"></a>
+## [2.0.9-unstable.2](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.9-unstable.1...aws-amplify-react-native@2.0.9-unstable.2) (2018-12-03)
+
+
+### Features
+
+* **@aws-amplify/react-native:** Add dial code selector to sign-up ([ab5efc3](https://github.com/aws/aws-amplify/commit/ab5efc3))
+
+
+
+
+<a name="2.0.9-unstable.1"></a>
+## [2.0.9-unstable.1](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.9-unstable.0...aws-amplify-react-native@2.0.9-unstable.1) (2018-11-29)
+
+
+### Bug Fixes
+
+* **aws-amplify-react-native:** render required attributes when requiring new passwords ([faa502f](https://github.com/aws/aws-amplify/commit/faa502f))
+
+
+
+
+<a name="2.0.9-unstable.0"></a>
+## [2.0.9-unstable.0](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.8...aws-amplify-react-native@2.0.9-unstable.0) (2018-11-23)
+
+
+### Bug Fixes
+
+* **aws-amplify-react:** add I18n in Greetings ([e549db7](https://github.com/aws/aws-amplify/commit/e549db7))
+
+
+
+
 <a name="2.0.8"></a>
 ## [2.0.8](https://github.com/aws/aws-amplify/compare/aws-amplify-react-native@2.0.8-unstable.0...aws-amplify-react-native@2.0.8) (2018-10-29)