+++ packages/aws-amplify-vue/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.8-unstable.4"></a>
+## [0.1.8-unstable.4](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.8-unstable.3...aws-amplify-vue@0.1.8-unstable.4) (2018-11-30)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.1.8-unstable.3"></a>
+## [0.1.8-unstable.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.8-unstable.2...aws-amplify-vue@0.1.8-unstable.3) (2018-11-30)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.1.8-unstable.2"></a>
+## [0.1.8-unstable.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.8-unstable.1...aws-amplify-vue@0.1.8-unstable.2) (2018-11-26)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.1.8-unstable.1"></a>
+## [0.1.8-unstable.1](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.8-unstable.0...aws-amplify-vue@0.1.8-unstable.1) (2018-11-21)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
+<a name="0.1.8-unstable.0"></a>
+## [0.1.8-unstable.0](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.7...aws-amplify-vue@0.1.8-unstable.0) (2018-11-19)
+
+
+
+
+**Note:** Version bump only for package aws-amplify-vue
+
 <a name="0.1.7"></a>
 ## [0.1.7](https://github.com/aws-amplify/amplify-js/compare/aws-amplify-vue@0.1.7-unstable.1...aws-amplify-vue@0.1.7) (2018-11-12)